iter6: fix

+++ migrations/20250818140935_add_table_orders.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 CREATE table IF NOT EXISTS orders
 (
-    id          int UNIQUE NOT NULL,
+    id          bigint UNIQUE NOT NULL,
     user_id     int        NOT NULL,
     status      text       NOT NULL,
     accrual     int        NOT NULL DEFAULT 0,
